+++ main.struct.go
@@ -6,7 +6,7 @@ import (
 )
 
 const (
-	version      = "2.2.0"
+	version      = "2.2.1"
 	constOK      = "ok"
 	updateString = "Update"
 	createString = "Create"